:up: VersionTest using traits.
- Move providers to a trait.
- Implement ``TestConstructorHelper`` and ``IComparableTestHelper``.
